+++ src/af/audioFeatureDSP/audioFeatureDSP.c
@@ -7,10 +7,10 @@
 //4 BAND EQ -- (per frame) -- FREQUENCY DOMAIN
 //Crest Factor (whole buffer) -- TIME DOMAIN
 
-double calculateSpectralCentroid(uint64_t onsetIndex)
+float calculateSpectralCentroid(uint64_t onsetIndex)
 {
-    double _spectralCentroid = 0.0;
-    double magnitudeSum = 0.0;
+    float _spectralCentroid = 0.0;
+    float magnitudeSum = 0.0;
     for (uint32_t i = 0; i < FFT_N2_LENGTH; i++)
     {
         _spectralCentroid += (magnitudeBeatBuffer[onsetIndex][i] * i);
@@ -23,13 +23,13 @@ double calculateSpectralCentroid(uint64_t onsetIndex)
     return (_spectralCentroid / FFT_N2_LENGTH) * sampleRate;
 }
 
-double calculateBandL(uint64_t onsetIndex) {
+float calculateBandL(uint64_t onsetIndex) {
     return (magnitudeBeatBuffer[onsetIndex][1] + magnitudeBeatBuffer[onsetIndex][2] ) / 2.0; //187Hz
 }
 //BIN * 96000 / 512 = freq
 //BIN * 48000 / 512 = freq
-double calculateBandML(uint64_t onsetIndex) {
-    double ret = 0;
+float calculateBandML(uint64_t onsetIndex) {
+    float ret = 0;
     for(int i = 3; i <= 10; i++)
     {
         ret += (magnitudeBeatBuffer[onsetIndex][i]); //937Hz
@@ -37,8 +37,8 @@ double calculateBandML(uint64_t onsetIndex) {
     return (ret / (11-3));
 }
 
-double calculateBandMH(uint64_t onsetIndex) {
-    double ret = 0;
+float calculateBandMH(uint64_t onsetIndex) {
+    float ret = 0;
     for(int i = 11; i <= 22; i++)
     {
         ret += (magnitudeBeatBuffer[onsetIndex][i]); //2 062Hz
@@ -46,8 +46,8 @@ double calculateBandMH(uint64_t onsetIndex) {
     return (ret / (23-11));
 }
 
-double calculateBandH(uint64_t onsetIndex) {
-    double ret = 0;
+float calculateBandH(uint64_t onsetIndex) {
+    float ret = 0;
     for(int i = 23; i <= 214; i++)
     {
         ret += (magnitudeBeatBuffer[onsetIndex][i]); //20 062Hz
@@ -55,9 +55,9 @@ double calculateBandH(uint64_t onsetIndex) {
     return (ret / (215-23));
 }
 
-double calculateCrestFactor(double* buffer, uint64_t audioIndex, uint64_t length) {
-    double peak = 0.0;
-    double rms = 0.0;
+float calculateCrestFactor(float* buffer, uint64_t audioIndex, uint64_t length) {
+    float peak = 0.0;
+    float rms = 0.0;
 
     // Find the peak value
     for (int i = 0; i < length; i++) {
