+++ src/af/transientDetectionDSP/EnvelopeFollowerPeakHoldAF.c
@@ -3,7 +3,7 @@
 #include "../../hal/globalDefinitions.hpp"
 #include <inttypes.h>
 
-void EnvelopeFollowerPeakHoldInitAf(double window_size_ms)
+void EnvelopeFollowerPeakHoldInitAf(float window_size_ms)
 {
     // prev_max = 0.0;
     BufferCircularInitAf((unsigned int)(window_size_ms / 1000.0 * slowsampleRate));
@@ -11,9 +11,9 @@ void EnvelopeFollowerPeakHoldInitAf(double window_size_ms)
 
 uint16_t blockCounterAf = 0;
 uint16_t blockThresholdAf = 8;
-double heldMaxAf = 0;
+float heldMaxAf = 0;
 
-double EnvelopeFollowerPeakHoldProcessAf(double x)
+float EnvelopeFollowerPeakHoldProcessAf(float x)
 {
     BufferCircularPush_sampleAf(x);
     if (blockCounterAf > blockThresholdAf)
