+++ src/af/transientDetectionDSP/envDeps/BufferCircularAF.c
@@ -1,14 +1,14 @@
 #include "BufferCircularAF.h"
 #include <memory.h>
 
-double __attribute__((section(".sdram_bss"))) EnvCircularBufferAf[ENVELOPE_FOLLOWER_BUFFER_SIZE];
+float __attribute__((section(".sdram_bss"))) EnvCircularBufferAf[ENVELOPE_FOLLOWER_BUFFER_SIZE];
 unsigned int index_writeAf;
 unsigned int len_bufferAf; // Must be nearest power of 2
 unsigned int wrap_maskAf;  // Must be (len_buffer - 1)
 
 void BufferCircularClearAf()
 {
-    memset(&EnvCircularBufferAf[0], 0, len_bufferAf * sizeof(double));
+    memset(&EnvCircularBufferAf[0], 0, len_bufferAf * sizeof(float));
 }
 
 void BufferCircularInitAf(unsigned int num_samples)
@@ -27,14 +27,14 @@ void BufferCircularInit_power_of_twoAf(unsigned int num_samples_power_of_two)
     BufferCircularClearAf();
 }
 
-void BufferCircularPush_sampleAf(double input)
+void BufferCircularPush_sampleAf(float input)
 {
     EnvCircularBufferAf[index_writeAf++] = input;
 
     index_writeAf &= wrap_maskAf;
 }
 
-double BufferCircularReadAf(int samples_delay)
+float BufferCircularReadAf(int samples_delay)
 {
     int index_read = (index_writeAf - 1) - samples_delay;
     index_read &= wrap_maskAf;
@@ -42,9 +42,9 @@ double BufferCircularReadAf(int samples_delay)
     return EnvCircularBufferAf[index_read];
 }
 
-double BufferCircularMaxAf()
+float BufferCircularMaxAf()
 {
-    double max = EnvCircularBufferAf[0];
+    float max = EnvCircularBufferAf[0];
     for (unsigned int i = 0; i < len_bufferAf; i++)
     {
         if (EnvCircularBufferAf[i] > max)
