+++ src/af/transientDetectionDSP/envDeps/BufferCircularAF.h
@@ -7,7 +7,7 @@ void BufferCircularInit_power_of_twoAf(unsigned int num_samples_power_of_two);
 
 void BufferCircularClearAf();
 
-void BufferCircularPush_sampleAf(double input);
-double BufferCircularReadAf(int samples_delay);
+void BufferCircularPush_sampleAf(float input);
+float BufferCircularReadAf(int samples_delay);
 
-double BufferCircularMaxAf();
+float BufferCircularMaxAf();
